@@ -37,14 +37,16 @@
   "devDependencies": {
     "babel-cli": "^6.18.0",
     "babel-core": "^6.24.0",
-    "babel-loader": "^6.2.10",
-    "babel-preset-es2015": "^6.18.0",
-    "babel-preset-react": "^6.16.0",
-    "babel-preset-stage-2": "^6.18.0",
-    "create-react-class": "^15.5.2",
-    "exports-loader": "^0.6.2",
-    "file-loader": "^0.10.1",
-    "highlight.js": "^9.10.0",
+    "babel-loader": "^7.1.2",
+    "babel-plugin-add-module-exports": "^0.2.1",
+    "babel-plugin-transform-export-default": "^7.0.0-alpha.20",
+    "babel-preset-env": "^1.6.1",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-stage-2": "^6.24.1",
+    "create-react-class": "^15.6.2",
+    "exports-loader": "^0.6.4",
+    "file-loader": "^1.1.6",
+    "highlight.js": "^9.12.0",
     "imports-loader": "^0.7.1",
     "jsdom": "^9.9.1",
     "mocha": "^3.2.0",
@@ -52,12 +54,13 @@
     "nightwatch": "^0.9.6",
     "prop-types": "^15.5.8",
     "raw-loader": "^0.5.1",
-    "react": "^15.5.4",
-    "react-dom": "^15.5.4",
-    "react-highlight": "^0.9.0",
-    "sinon": "^2.0.0",
-    "webpack": "^2.2.1",
-    "webpack-dev-server": "^2.4.2"
+    "react": "^16.2.0",
+    "react-dom": "^16.2.0",
+    "react-highlight": "^0.10.0",
+    "sinon": "^4.1.3",
+    "uglifyjs-webpack-plugin": "^1.1.6",
+    "webpack": "^3.10.0",
+    "webpack-dev-server": "^2.9.6"
   },
   "dependencies": {
     "highcharts": "^6.0.0"
