@@ -1,1 +1,4 @@
-module.exports = require('./chartsFactory.jsx')('StockChart', require("highcharts/highstock"));
+import chartsFactory from './chartsFactory.jsx';
+import highstock from 'highcharts/highstock';
+
+export default chartsFactory('StockChart', highstock);
