@@ -1,1 +1,4 @@
-module.exports = require('./chartsFactory.jsx')('Chart', require("highcharts"));
+import chartsFactory from './chartsFactory.jsx';
+import highcharts from 'highcharts';
+
+export default chartsFactory('Chart', highcharts);
